refactor: move EmailValidatorAdapter and PasswordValidatorAdapter to main layer
