+++ src/main/adapters/express/express-route-adapter.ts
@@ -1,4 +1,4 @@
-import { type Controller, type HttpRequest, type HttpResponse } from '../../presentation/protocols'
+import { type Controller, type HttpRequest, type HttpResponse } from '../../../presentation/protocols'
 import { type Request, type Response } from 'express'
 
 /**
@@ -7,7 +7,7 @@ import { type Request, type Response } from 'express'
  * nest caso o formato HttpRequest do controller deve ser modificado para se enquadrar
  * no formato Request do express
  */
-export function adaptRoute (controller: Controller) {
+export function expressRouteAdapter (controller: Controller) {
   return async (req: Request, res: Response) => {
     const httpRequest: HttpRequest = {
       body: req.body
