+++ src/main/adapters/express/index.ts
@@ -1,0 +1,1 @@
+export * from './express-route-adapter'
