+++ src/main/adapters/validator/email-validator-adapter.ts
@@ -1,4 +1,4 @@
-import { type EmailValidator } from '../presentation/protocols'
+import { type EmailValidator } from '../../../presentation/protocols'
 import validator from 'validator'
 
 export class EmailValidatorAdapter implements EmailValidator {
