+++ src/main/adapters/validator/index.ts
@@ -1,0 +1,2 @@
+export * from './email-validator-adapter'
+export * from './password-validator-adapter'
