+++ src/main/adapters/validator/password-validator-adapter.ts
@@ -1,4 +1,4 @@
-import { type PasswordValidator } from '../presentation/protocols'
+import { type PasswordValidator } from '../../../presentation/protocols'
 import validator from 'validator'
 
 export class PasswordValidatorAdapter implements PasswordValidator {
