+++ src/main/factories/signup/signup-validation-factory.ts
@@ -6,8 +6,8 @@ import {
   PasswordValidation
 } from '../../../presentation/helpers/validators/validations'
 import { type Validation } from '../../../presentation/controllers/signup/signup-controller-protocols'
-import { EmailValidatorAdapter } from '../../../utils/email-validator-adapter'
-import { PasswordValidatorAdapter } from '../../../utils/password-validator-adapter'
+import { EmailValidatorAdapter } from '../../adapters/validator/email-validator-adapter'
+import { PasswordValidatorAdapter } from '../../adapters/validator/password-validator-adapter'
 
 export function makeSignUpValidation (): ValidationComposite {
   const validations: Validation[] = []
