+++ src/main/routes/signup.ts
@@ -1,7 +1,7 @@
 import { type Router } from 'express'
 import { makeSignUpController } from '../factories/signup/signup-factory'
-import { adaptRoute } from '../adapters/express-route-adapter'
+import { expressRouteAdapter } from '../adapters/express'
 
 export default function (router: Router): void {
-  router.post('/signup', adaptRoute(makeSignUpController()))
+  router.post('/signup', expressRouteAdapter(makeSignUpController()))
 }
